pyproject.toml
[project]
name = "wikidiary"
version = "0.1.0"
description = "Interacting with diary hosted on semantic mediawiki"
authors = [
    {name = "Malte Sahrhage"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "boto3 (>=1.37.13,<2.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "pillow (>=11.1.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

